requirements.txt
python-telegram-bot==21.3
anthropic>=0.40.0
httpx>=0.27.0
aiohttp>=3.9.0
feedparser>=6.0.11
yfinance>=0.2.40
python-dotenv>=1.0.1
apscheduler>=3.10.0
